Refactors admin payments page
Improves the admin payments page by removing the separate subscription statistics tab and integrating its relevant summary data into the main transactions list.

This change streamlines the UI and provides a more cohesive overview of payment activity.

Additionally, this change moves the transaction summary and filters out of the TransactionsList component.
It was decided that these components were not a part of the table.

Updates the RevenueAnalyticsChart component to include a background color for the Popover component.